feat: implement `to_rgb` and `to_hex_triplet` for white enum
- refactor: move White enum and related implementations to separate module
